Remove Linux builds
Building on the Travis Linux environment is basically impossible right now as Travis refuses to use the correct standard library for the provided compiler, meaning all the Linux builds fail. This seems to be a common problem (https://github.com/travis-ci/travis-ci/issues/6300), so we should keep an eye out for changes and hopefully add the Linux builds back soon.