Fix spanning allele assignments
Ensure all spanning alleles (*) are included in haplotypes by appending to end of current defined region. This will mean that the resolved allele of a VCF record may not overlap with the record itself, and that multiple records map to the same allele, but it means that testing for inclusion is straightforward.