Force HaplotypeGenerator to merge block adjacent to lhs of indel
A reoccurring bug in Octopus is when an indel call is not haplotyped with a site adjacent to it - where the REF pad in VCF goes. This can cause various problems when converting to VCF. For example if the calls with the indel are a different ploidy to the calls before the indel.