Avoid deadlocks with ThreadPool::try_push
Since we use the ThreadPool to run calling tasks, if we then try to push to the pool from once of these tasks (e.g. during candidate generation), we can end up in deadlock if the pool is full. To avoid this, we must always keep the option of running in the current thread of execution, if the pool is full.